pyproject.toml
# ==========================================================
#                  CORRECTED pyproject.toml
# ==========================================================

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulso-back"
version = "1.0.0"
description = "API + ETL Backend para Dashboard Cobranzas Telefónica"
authors = [{name = "Ricky", email = "ricky@telefonica.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT License" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]
# --- DEPENDENCIAS PRINCIPALES (PARA PRODUCCIÓN) ---
dependencies = [
    # FastAPI & Web Framework
    "fastapi>=0.115.14",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.19", # Necesario para subida de archivos en FastAPI
    # Database & Cache
    # "psycopg2-binary>=2.9.9", # Replaced by asyncpg
    "redis>=5.0.1",
    "sqlalchemy>=2.0.23", # Keep for now, may be removed if not used by anything else
    "alembic>=1.13.1", # Keep for now, may be removed if not used by anything else
    # Google Cloud
    "google-cloud-bigquery>=3.13.0",
    "google-cloud-storage>=2.10.0",
    "google-auth>=2.23.4",
    # Data Processing & Config
    "pandas>=2.1.3",
    "numpy>=1.25.2",
    "pydantic-settings>=2.1.0",
    # ETL & Scheduling
    "celery[redis]>=5.3.4", # Especificamos 'redis' como backend
    "apscheduler>=3.10.4",
    # Monitoring & Logging
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "tenacity>=8.2.3",
    "asyncpg>=0.30.0",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
    "starlette>=0.40.0",
    "pydantic>=2.5.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=24.3.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

# --- DEPENDENCIAS OPCIONALES (PARA DESARROLLO) ---
[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=24.3.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

# --- SCRIPT DE EJECUCIÓN ---
[project.scripts]
dev = "app.main:main"


# --- ADD THIS SECTION ---
# This tells setuptools to only look for the 'app' package,
# resolving the "Multiple top-level packages" error.
[tool.setuptools]
packages = ["app"]


[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
